Test add capacity with restart node, when capacity near full ratio (#1996)
* Test add capacity with restart node, when capacity near full ratio

- Use monitoring to fill up the cluster, and create a new fixture in 'monitoring/conftest'
that will help with storageutilization of rbd, and cephfs.
- Perform IO using the method 'run_io' of the pod object.
- Make use in the functions that written in the branch 'WIP-clexp-entry'.
After this branch will be merged into master I will use them directly.
- Start one or more of the worker nodes when add_capacity is executed.


* add new conftest file in z_cluster


* delete few checks in the test


* Delete 2 helper functions

Delete 2 functions 'get_percent_used_capacity', 'check_pods_in_running_state'.
We will add these function as soon as branch 'wip-add-capacity-e_e' will be merged to master.
We don't want to have duplicate functions.


* Unite the two pytest parameters into one


* Edit 'workload_storageutilization_rbd', 'workload_storageutilization_cephfs' fixures

- Use f-string instead of format
- delete 'pvc_factory' from the test
- Add docstring in 'workload_fio_storageutilization'


* make sure 'request' parameter is provided


* Correct spelling mistakes


* move 'measurement_dir' fixture to 'tests/conftest' dir


* Change the condition of 'delete_fio_data_after_test'

change the condition of 'delete_fio_data_after_test' so it will
not obstructs the original meaning.


* Simplified the passing arguments in 'workload_fio_storageutilization'

Don't pass the 'request' object to 'workload_fio_storageutilization'.
The test namespace will be deleted in the teardown part, without adding
finalizer.


* make it more clear why and when the data are actually deleted


* Mark the test as 'tier4a' instead of 'tier4'


* increasing the timeout

increasing the timeout when waiting for osd pods,
and when waiting for ceph health check to be ok.

+++ ocs_ci/ocs/fiojob.py
@@ -299,6 +299,8 @@ def workload_fio_storageutilization(
     target_percentage=None,
     target_size=None,
     with_checksum=False,
+    keep_fio_data=False,
+    minimal_time=480,
 ):
     """
     This function implements core functionality of fio storage utilization
@@ -339,6 +341,10 @@ def workload_fio_storageutilization(
             fio is stored on the volume, and reclaim policy of the volume is
             changed to ``Retain`` so that the volume is not removed during test
             teardown for later verification runs
+        keep_fio_data (bool): If true, keep the fio data after the fio
+            storage utilization is completed. Else if false, deletes the fio data.
+        minimal_time (int): Minimal number of seconds to monitor a system.
+            (See more details in the function 'measure_operation')
 
     Returns:
         dict: measurement results with timestamps and other medatada from
@@ -486,7 +492,7 @@ def workload_fio_storageutilization(
             fio_job_file, write_timeout, pvc_size, target_percentage),
         test_file,
         measure_after=True,
-        minimal_time=480)
+        minimal_time=minimal_time)
 
     # we don't need to delete anything if this fixture has been already
     # executed
@@ -548,6 +554,9 @@ def workload_fio_storageutilization(
         # Without checksum, we just need to make sure that data were deleted
         # and wait for this to happen to avoid conflicts with tests executed
         # right after this one.
-        delete_fio_data(fio_job_file, is_storage_reclaimed)
+        if not keep_fio_data:
+            delete_fio_data(fio_job_file, is_storage_reclaimed)
+        else:
+            logging.info("The fio data will be deleted during project teardown")
 
     return measured_op
